requirements.txt
Flask==3.1.1
pycryptodome==3.10.1
ttkbootstrap==1.10.1
qrcode==7.3.1
Pillow==9.5.0 